Use double-click when user want to deselect Character or Object
User don't want to deselect Character or Object when he's editing. So I changed single click to double click when user want to deselect. Therefore he would never lose the target whenever he control the camera.

Moreover, the middle and right button are blocked. So the middle and right button will not select or move Character when user control camera.